restructured module
