+++ tests/test_cls_helper_base.py
@@ -4,7 +4,8 @@ try:
 except:
     pass
 import unittest
-from src.kwargs_util import HelperBase
+from src.helper.base import HelperBase
+
 
 
 class MyHelper(HelperBase):
